+++ eras/mary/impl/src/Cardano/Ledger/Mary/Core.hs
@@ -1,6 +1,3 @@
-{-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE UndecidableSuperClasses #-}
-
 module Cardano.Ledger.Mary.Core (
   MaryEraTxBody (..),
   module Cardano.Ledger.Allegra.Core,
@@ -8,13 +5,4 @@ module Cardano.Ledger.Mary.Core (
 where
 
 import Cardano.Ledger.Allegra.Core
-import Cardano.Ledger.Mary.Value (MultiAsset (..))
-import Data.Set (Set)
-import Lens.Micro (Lens', SimpleGetter)
-
-class AllegraEraTxBody era => MaryEraTxBody era where
-  mintTxBodyL :: Lens' (TxBody era) (MultiAsset (EraCrypto era))
-
-  mintValueTxBodyF :: SimpleGetter (TxBody era) (Value era)
-
-  mintedTxBodyF :: SimpleGetter (TxBody era) (Set (ScriptHash (EraCrypto era)))
+import Cardano.Ledger.Mary.TxBody (MaryEraTxBody (..))